+++ backend/src/utils/network/xtream.rs
@@ -5,15 +5,13 @@ use crate::model::{is_input_expired, xtream_mapping_option_from_target_options, 
 use crate::model::{InputSource, ProxyUserCredentials};
 use crate::processing::parser::xtream;
 use crate::processing::parser::xtream::parse_xtream_series_info;
-use crate::repository::{needs_update_info_details, BPlusTreeQuery, BPlusTreeUpdate, FlushPolicy};
 use crate::repository::VirtualIdRecord;
-use crate::repository::{ensure_input_storage_path, get_input_storage_path, get_target_storage_path};
-use crate::repository::{get_live_cat_collection_path, get_series_cat_collection_path, get_vod_cat_collection_path, xtream_get_file_path, CategoryEntry};
+use crate::repository::{get_input_storage_path, get_target_storage_path};
 use crate::repository::{get_target_id_mapping, rewrite_provider_series_info_episode_virtual_id, ProviderEpisodeKey};
-use crate::repository::{persist_input_vod_info, persists_input_series_info, write_playlist_batch_item_upsert, write_playlist_item_update};
+use crate::repository::{persist_input_xtream_playlist_cluster_to_disk, persist_input_vod_info, persists_input_series_info, write_playlist_batch_item_upsert, write_playlist_item_update};
 use crate::utils::request;
 use chrono::{DateTime, Utc};
-use log::{debug, error, info, warn};
+use log::{error, info, warn};
 use shared::error::TuliproxError;
 use shared::model::{PlaylistEntry, PlaylistGroup, ProxyUserStatus, SeriesStreamProperties,
                     StreamProperties, VideoStreamProperties, XtreamCluster, XtreamPlaylistItem,
@@ -21,16 +19,11 @@ use shared::model::{PlaylistEntry, PlaylistGroup, ProxyUserStatus, SeriesStreamP
 use shared::utils::{extract_extension_from_url, get_i64_from_serde_value, get_string_from_serde_value, sanitize_sensitive_info, Internable, PROVIDER_SCHEME_PREFIX};
 use std::collections::HashMap;
 use std::io::Error;
-use std::path::Path;
 use std::str::FromStr;
 
-use crate::model::XtreamCategory;
-use crate::utils::request::DynReader;
-use fs2::FileExt;
-use std::fs::File;
 use std::sync::Arc;
 
-use shared::{info_err, notify_err, notify_err_res};
+use shared::info_err;
 
 const THREE_DAYS_IN_SECS: i64 = 3 * 24 * 60 * 60;
 
@@ -404,11 +397,11 @@ pub async fn download_xtream_playlist(app_config: &Arc<AppConfig>, client: &reqw
                 (Ok(category_content), Ok(stream_content)) => {
                     if cfg.disk_based_processing {
                         // trace!("Using disk input playlist optimization for cluster {}", xtream_cluster);
-                        if let Err(err) = process_xtream_cluster_to_disk(app_config, input, *xtream_cluster, category_content, stream_content).await {
-                            error!("process_xtream_cluster_to_disk failed: {err}");
+                        if let Err(err) = persist_input_xtream_playlist_cluster_to_disk(app_config, input, *xtream_cluster, category_content, stream_content).await {
+                            error!("persist_input_xtream_playlist_cluster_to_disk failed: {err}");
                             errors.push(err);
-                        } else {
-                            // trace!("process_xtream_cluster_to_disk succeeded for cluster {}", xtream_cluster);
+                        //} else {
+                            // trace!("persist_input_xtream_playlist_cluster_to_disk succeeded for cluster {}", xtream_cluster);
                         }
                     } else {
                         // trace!("Using in-memory playlist parsing for cluster {}", xtream_cluster);
@@ -516,487 +509,4 @@ pub fn create_vod_info_from_item(target: &ConfigTarget, user: &ProxyUserCredenti
     doc.movie_data.custom_sid = None;
 
     serde_json::to_string(&doc).unwrap_or(String::new())
-}
-
-const BATCH_SIZE: usize = 1000;
-
-fn preserve_details_for_disk_cluster(
-    old_path: &Path,
-    tmp_path: &Path,
-) -> Result<(), TuliproxError> {
-    let mut old_props_by_provider_id: HashMap<u32, StreamProperties> = HashMap::new();
-
-    if old_path.exists() {
-        if let Ok(mut old_query) = BPlusTreeQuery::<u32, XtreamPlaylistItem>::try_new(old_path) {
-            for (_, item) in old_query.iter() {
-                if let Some(props) = item.additional_properties.as_ref() {
-                    let should_preserve = props.has_details()
-                        || matches!(
-                            props,
-                            StreamProperties::Live(live)
-                                if live.video.is_some()
-                                    || live.audio.is_some()
-                                    || live.last_probed_timestamp.is_some()
-                                    || live.last_success_timestamp.is_some()
-                        );
-                    if should_preserve {
-                        old_props_by_provider_id.insert(item.provider_id, props.clone());
-                    }
-                }
-            }
-        }
-    }
-
-    if old_props_by_provider_id.is_empty() {
-        return Ok(());
-    }
-
-    let mut updates: Vec<(u32, XtreamPlaylistItem)> = Vec::new();
-    if let Ok(mut new_query) = BPlusTreeQuery::<u32, XtreamPlaylistItem>::try_new(tmp_path) {
-        for (_, mut item) in new_query.iter() {
-            let Some(new_props) = item.additional_properties.as_mut() else {
-                continue;
-            };
-
-            let Some(old_props) = old_props_by_provider_id.get(&item.provider_id) else {
-                continue;
-            };
-
-            let preserve_info_details =
-                old_props.has_details() && !needs_update_info_details(new_props, old_props);
-
-            let mut changed = false;
-            match (new_props, old_props) {
-                (StreamProperties::Video(v_new), StreamProperties::Video(v_old))
-                    if preserve_info_details && v_old.details.is_some() =>
-                {
-                    if v_new.details != v_old.details {
-                        v_new.details.clone_from(&v_old.details);
-                        changed = true;
-                    }
-                }
-                (StreamProperties::Series(s_new), StreamProperties::Series(s_old))
-                    if preserve_info_details && s_old.details.is_some() =>
-                {
-                    if s_new.details != s_old.details {
-                        s_new.details.clone_from(&s_old.details);
-                        changed = true;
-                    }
-                }
-                (StreamProperties::Live(l_new), StreamProperties::Live(l_old)) => {
-                    // Preserve live probe state across rebuilds so periodic probe intervals remain stable.
-                    let prev_video = l_new.video.clone();
-                    let prev_audio = l_new.audio.clone();
-                    let prev_last_probed = l_new.last_probed_timestamp;
-                    let prev_last_success = l_new.last_success_timestamp;
-
-                    if l_new.video.is_none() {
-                        l_new.video.clone_from(&l_old.video);
-                    }
-                    if l_new.audio.is_none() {
-                        l_new.audio.clone_from(&l_old.audio);
-                    }
-                    if l_new.last_probed_timestamp.is_none() {
-                        l_new.last_probed_timestamp = l_old.last_probed_timestamp;
-                    }
-                    if l_new.last_success_timestamp.is_none() {
-                        l_new.last_success_timestamp = l_old.last_success_timestamp;
-                    }
-
-                    changed = l_new.video != prev_video
-                        || l_new.audio != prev_audio
-                        || l_new.last_probed_timestamp != prev_last_probed
-                        || l_new.last_success_timestamp != prev_last_success;
-                }
-                _ => {}
-            }
-
-            if changed {
-                updates.push((item.provider_id, item));
-            }
-        }
-    }
-
-    if !updates.is_empty() {
-        let mut tree = BPlusTreeUpdate::<u32, XtreamPlaylistItem>::try_new_with_backoff(tmp_path)
-            .map_err(|e| notify_err!("Failed to open tmp tree for merge: {e}"))?;
-        let refs: Vec<(&u32, &XtreamPlaylistItem)> =
-            updates.iter().map(|(id, pli)| (id, pli)).collect();
-        tree.update_batch(&refs)
-            .map_err(|e| notify_err!("Failed to update tmp tree during merge: {e}"))?;
-        tree.commit()
-            .map_err(|e| notify_err!("Failed to commit tmp tree merge: {e}"))?;
-    }
-
-    Ok(())
-}
-
-#[allow(clippy::too_many_lines)]
-async fn process_xtream_cluster_to_disk(
-    app_config: &Arc<AppConfig>,
-    input: &ConfigInput,
-    cluster: XtreamCluster,
-    categories: DynReader,
-    streams: DynReader,
-) -> Result<(), TuliproxError> {
-    let cfg = app_config.config.load();
-    let storage_path = ensure_input_storage_path(&cfg, &input.name).await?;
-    let xtream_path = xtream_get_file_path(&storage_path, cluster);
-
-    // Channel for transferring items from Parser (Async Task) to Consumer (Blocking Task)
-    let (tx, mut rx) = tokio::sync::mpsc::channel::<XtreamPlaylistItem>(BATCH_SIZE * 2);
-    let input_clone = input.clone();
-
-    // 1. Parser Task: Runs the async parsing logic
-    // We move the readers into this task.
-    let parse_task = tokio::spawn(async move {
-        let tx_for_closure = tx.clone();
-        let res = xtream::parse_xtream_streaming(&input_clone, cluster, categories, streams, move |item| {
-            // FIX: Copy/Clone needed data BEFORE moving the 'item' into the channel
-            let item_id = item.virtual_id;
-
-            // We use blocking_send because the closure provided by the parser library is synchronous.
-            // This is safe here because it runs within its own tokio::spawn task.
-            if let Err(e) = tx_for_closure.blocking_send(item) {
-                error!("Channel closed while processing {cluster} for item {item_id}: {e}");
-                return notify_err_res!("Channel closed while processing {cluster}");
-            }
-            Ok(())
-        }).await;
-
-        // CRITICAL: Explicitly drop the sender to signal rx.blocking_recv() to stop.
-        // This prevents the consumer from waiting forever if the parser fails.
-        drop(tx);
-        res
-    });
-
-    // 2. Consumer Task: Handles heavy Disk I/O (BPlusTree updates)
-    let xtream_path_for_consumer = xtream_path.clone();
-    let consumer_task = tokio::task::spawn_blocking(move || {
-        let tmp_xtream_path = xtream_path_for_consumer.with_extension("tmp");
-
-        crate::repository::BPlusTree::<u32, XtreamPlaylistItem>::new()
-            .store(&tmp_xtream_path)
-            .map_err(|e| {
-                error!("Failed to initialize ghost BPlusTree at {}: {e}", tmp_xtream_path.display());
-                notify_err!("Init tree error {e}")
-            })?;
-
-        let mut tree: BPlusTreeUpdate<u32, XtreamPlaylistItem> = BPlusTreeUpdate::try_new_with_backoff(&tmp_xtream_path)
-            .map_err(|e| {
-                error!("Failed to open ghost tree at {}: {e}", tmp_xtream_path.display());
-                notify_err!("Failed to open tree {e}")
-            })?;
-        tree.set_flush_policy(FlushPolicy::Batch);
-
-        let mut buffer = Vec::with_capacity(BATCH_SIZE);
-
-        // This loop exits when all 'tx' clones are dropped (signaling end of stream)
-        while let Some(item) = rx.blocking_recv() {
-            buffer.push(item);
-            if buffer.len() >= BATCH_SIZE {
-                let batch: Vec<(&u32, &XtreamPlaylistItem)> = buffer.iter().map(|i| (&i.provider_id, i)).collect();
-                let prepared = BPlusTreeUpdate::<u32, XtreamPlaylistItem>::prepare_upsert_batch(&batch).map_err(|e| {
-                    error!("Batch prepare failed for cluster {cluster}: {e}");
-                    notify_err!("Prepare failed {e}")
-                })?;
-                tree.upsert_batch_encoded(prepared).map_err(|e| {
-                    error!("Batch upsert failed for cluster {cluster}: {e}");
-                    notify_err!("Upsert failed {e}")
-                })?;
-                buffer.clear();
-            }
-        }
-
-        // Final batch processing
-        if !buffer.is_empty() {
-            let batch: Vec<(&u32, &XtreamPlaylistItem)> = buffer.iter().map(|i| (&i.provider_id, i)).collect();
-            let prepared = BPlusTreeUpdate::<u32, XtreamPlaylistItem>::prepare_upsert_batch(&batch).map_err(|e| {
-                error!("Final batch prepare failed for cluster {cluster}: {e}");
-                notify_err!("Prepare failed {e}")
-            })?;
-            tree.upsert_batch_encoded(prepared).map_err(|e| {
-                error!("Final batch upsert failed for cluster {cluster}: {e}");
-                notify_err!("Upsert failed {e}")
-            })?;
-        }
-
-        tree.commit().map_err(|e| {
-            error!("Commit failed for cluster {cluster}: {e}");
-            notify_err!("Commit failed {e}")
-        })?;
-        Ok::<(), TuliproxError>(())
-    });
-
-    // 3. Robust Joining of both tasks
-    // try_join! returns immediately if any task returns an error or panics.
-    let (parse_res, consumer_res) = tokio::try_join!(parse_task, consumer_task)
-        .map_err(|e| notify_err!("Task join error during cluster {cluster} update: {e}"))?;
-
-    // Handle internal errors from the tasks
-    let parsed_categories = parse_res?;
-    consumer_res?;
-
-    // --- Post-Processing & Atomic Swap ---
-
-    let col_path = match cluster {
-        XtreamCluster::Live => get_live_cat_collection_path(&storage_path),
-        XtreamCluster::Video => get_vod_cat_collection_path(&storage_path),
-        XtreamCluster::Series => get_series_cat_collection_path(&storage_path),
-    };
-
-    let tmp_col_path = col_path.with_extension("tmp");
-
-    // Use the parsed_categories returned from the task, not the 'categories' reader
-    save_xtream_categories_to_file(&tmp_col_path, &parsed_categories).await?;
-
-    let tmp_xtream_path = xtream_path.with_extension("tmp");
-
-    // Acquire file lock to ensure atomic operations
-    let swap_lock = app_config.file_locks.write_lock(&xtream_path).await;
-
-    {
-        let old_path = xtream_path.clone();
-        let new_tmp_path = tmp_xtream_path.clone();
-        tokio::task::spawn_blocking(move || preserve_details_for_disk_cluster(&old_path, &new_tmp_path))
-            .await
-            .map_err(|e| notify_err!("Merge task join error during cluster {cluster} update: {e}"))??;
-    }
-
-    // Optional compaction to optimize the newly created database file.
-    // Run on blocking pool because B+Tree lock backoff uses std::thread::sleep.
-    let compact_path = tmp_xtream_path.clone();
-    match tokio::task::spawn_blocking(move || -> Result<(), std::io::Error> {
-        if let Ok(mut tree_update) = BPlusTreeUpdate::<u32, XtreamPlaylistItem>::try_new_with_backoff(&compact_path) {
-            tree_update.compact(&compact_path)?;
-        }
-        Ok(())
-    }).await {
-        Ok(Ok(())) => {}
-        Ok(Err(e)) => {
-            error!("Failed to compact temporary database for {cluster} at {:?}: {e}", tmp_xtream_path.display());
-        }
-        Err(e) => {
-            error!("Compaction task join failed for {cluster} at {:?}: {e}", tmp_xtream_path.display());
-        }
-    }
-
-    // Atomic Swap: Replace old database with new one
-    if let Err(e) = crate::utils::rename_or_copy(&tmp_xtream_path, &xtream_path, false) {
-        error!("Failed to swap xtream database for {cluster} (from {:?} to {:?}): {e}", tmp_xtream_path.display(), xtream_path.display());
-        return notify_err_res!("Failed to swap database: {e}");
-    }
-
-    // Atomic Swap: Replace old categories with new ones
-    if let Err(e) = crate::utils::rename_or_copy(&tmp_col_path, &col_path, false) {
-        error!("Failed to swap xtream categories for {cluster} (from {:?} to {:?}): {e}", tmp_col_path.display(), col_path.display());
-        return notify_err_res!("Failed to swap categories: {e}");
-    }
-
-    // Cleanup: Remove temporary files if they still exist (defensive)
-    let _ = tokio::fs::remove_file(&tmp_xtream_path).await;
-    let _ = tokio::fs::remove_file(&tmp_col_path).await;
-
-    // Explicitly release the lock (though RAII would handle it too)
-    drop(swap_lock);
-
-    debug!("Cluster {cluster} updated and swapped successfully.");
-    Ok(())
-}
-
-async fn save_xtream_categories_to_file(col_path: &Path, categories: &[XtreamCategory]) -> Result<(), TuliproxError> {
-    let col_path_buf = col_path.to_path_buf();
-    let cat_entries: Vec<CategoryEntry> = categories.iter().map(|c| CategoryEntry {
-        category_id: c.category_id,
-        category_name: c.category_name.clone(),
-        parent_id: 0,
-    }).collect();
-
-    tokio::task::spawn_blocking(move || {
-        if let Ok(file) = File::create(&col_path_buf) {
-            if let Err(e) = file.lock_exclusive() {
-                warn!("Could not acquire exclusive lock for {}: {e}, proceeding without lock", col_path_buf.display());
-            }
-            serde_json::to_writer(&file, &cat_entries).map_err(|e| {
-                error!("Failed to write categories to file {}: {e}", col_path_buf.display());
-                notify_err!("Write failed: {e}")
-            })?;
-            let _ = file.unlock();
-        } else {
-            return notify_err_res!("Failed to create category file {}", col_path_buf.display());
-        }
-        Ok(())
-    }).await.map_err(|e| notify_err!("Spawn error {e}"))?
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use shared::model::LiveStreamProperties;
-    use shared::utils::Internable;
-    use tempfile::tempdir;
-
-    fn make_live_item(
-        provider_id: u32,
-        video: Option<&str>,
-        audio: Option<&str>,
-        last_probed_timestamp: Option<i64>,
-        last_success_timestamp: Option<i64>,
-    ) -> XtreamPlaylistItem {
-        XtreamPlaylistItem {
-            virtual_id: provider_id,
-            provider_id,
-            name: "live".intern(),
-            logo: "".intern(),
-            logo_small: "".intern(),
-            group: "group".intern(),
-            title: "".intern(),
-            parent_code: "".intern(),
-            rec: "".intern(),
-            url: "http://example.com/live.ts".intern(),
-            epg_channel_id: None,
-            xtream_cluster: XtreamCluster::Live,
-            additional_properties: Some(StreamProperties::Live(Box::new(LiveStreamProperties {
-                video: video.map(Internable::intern),
-                audio: audio.map(Internable::intern),
-                last_probed_timestamp,
-                last_success_timestamp,
-                ..Default::default()
-            }))),
-            item_type: shared::model::PlaylistItemType::Live,
-            category_id: 1,
-            input_name: "input_a".intern(),
-            channel_no: 0,
-            source_ordinal: 0,
-        }
-    }
-
-    fn write_single_item(path: &Path, item: &XtreamPlaylistItem) {
-        crate::repository::BPlusTree::<u32, XtreamPlaylistItem>::new()
-            .store(path)
-            .expect("tree creation should succeed");
-        let mut tree = BPlusTreeUpdate::<u32, XtreamPlaylistItem>::try_new_with_backoff(path)
-            .expect("tree open should succeed");
-        let batch: Vec<(&u32, &XtreamPlaylistItem)> = vec![(&item.provider_id, item)];
-        let prepared = BPlusTreeUpdate::<u32, XtreamPlaylistItem>::prepare_upsert_batch(&batch)
-            .expect("batch preparation should succeed");
-        tree.upsert_batch_encoded(prepared)
-            .expect("batch upsert should succeed");
-        tree.commit().expect("tree commit should succeed");
-    }
-
-    fn read_live_props(path: &Path, provider_id: u32) -> LiveStreamProperties {
-        let mut query =
-            BPlusTreeQuery::<u32, XtreamPlaylistItem>::try_new(path).expect("query open should succeed");
-        let item = query
-            .query_zero_copy(&provider_id)
-            .expect("query should succeed")
-            .expect("item should exist");
-        match item.additional_properties {
-            Some(StreamProperties::Live(live)) => *live,
-            other => panic!("expected live stream properties, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn preserve_details_for_disk_cluster_copies_missing_live_probe_fields() {
-        let dir = tempdir().expect("temp dir should be created");
-        let old_path = dir.path().join("old_live.db");
-        let tmp_path = dir.path().join("tmp_live.db");
-        let provider_id = 100_u32;
-
-        write_single_item(
-            &old_path,
-            &make_live_item(
-                provider_id,
-                Some("{\"codec_name\":\"h264\"}"),
-                Some("{\"codec_name\":\"aac\"}"),
-                Some(1_700_000_000),
-                Some(1_700_000_100),
-            ),
-        );
-        write_single_item(&tmp_path, &make_live_item(provider_id, None, None, None, None));
-
-        preserve_details_for_disk_cluster(&old_path, &tmp_path).expect("merge should succeed");
-
-        let merged = read_live_props(&tmp_path, provider_id);
-        assert_eq!(merged.video, Some("{\"codec_name\":\"h264\"}".intern()));
-        assert_eq!(merged.audio, Some("{\"codec_name\":\"aac\"}".intern()));
-        assert_eq!(merged.last_probed_timestamp, Some(1_700_000_000));
-        assert_eq!(merged.last_success_timestamp, Some(1_700_000_100));
-    }
-
-    #[test]
-    fn preserve_details_for_disk_cluster_does_not_override_existing_live_probe_fields() {
-        let dir = tempdir().expect("temp dir should be created");
-        let old_path = dir.path().join("old_live_existing.db");
-        let tmp_path = dir.path().join("tmp_live_existing.db");
-        let provider_id = 200_u32;
-
-        write_single_item(
-            &old_path,
-            &make_live_item(
-                provider_id,
-                Some("{\"codec_name\":\"h264\"}"),
-                Some("{\"codec_name\":\"aac\"}"),
-                Some(1_700_000_000),
-                Some(1_700_000_100),
-            ),
-        );
-        write_single_item(
-            &tmp_path,
-            &make_live_item(
-                provider_id,
-                Some("{\"codec_name\":\"hevc\"}"),
-                Some("{\"codec_name\":\"ac3\"}"),
-                Some(1_800_000_000),
-                Some(1_800_000_100),
-            ),
-        );
-
-        preserve_details_for_disk_cluster(&old_path, &tmp_path).expect("merge should succeed");
-
-        let merged = read_live_props(&tmp_path, provider_id);
-        assert_eq!(merged.video, Some("{\"codec_name\":\"hevc\"}".intern()));
-        assert_eq!(merged.audio, Some("{\"codec_name\":\"ac3\"}".intern()));
-        assert_eq!(merged.last_probed_timestamp, Some(1_800_000_000));
-        assert_eq!(merged.last_success_timestamp, Some(1_800_000_100));
-    }
-
-    #[test]
-    fn preserve_details_for_disk_cluster_fills_only_missing_live_probe_fields() {
-        let dir = tempdir().expect("temp dir should be created");
-        let old_path = dir.path().join("old_live_partial.db");
-        let tmp_path = dir.path().join("tmp_live_partial.db");
-        let provider_id = 300_u32;
-
-        write_single_item(
-            &old_path,
-            &make_live_item(
-                provider_id,
-                Some("{\"codec_name\":\"h264\"}"),
-                Some("{\"codec_name\":\"aac\"}"),
-                Some(1_700_000_000),
-                Some(1_700_000_100),
-            ),
-        );
-        write_single_item(
-            &tmp_path,
-            &make_live_item(
-                provider_id,
-                Some("{\"codec_name\":\"hevc\"}"),
-                None,
-                Some(1_800_000_000),
-                None,
-            ),
-        );
-
-        preserve_details_for_disk_cluster(&old_path, &tmp_path).expect("merge should succeed");
-
-        let merged = read_live_props(&tmp_path, provider_id);
-        assert_eq!(merged.video, Some("{\"codec_name\":\"hevc\"}".intern()));
-        assert_eq!(merged.audio, Some("{\"codec_name\":\"aac\"}".intern()));
-        assert_eq!(merged.last_probed_timestamp, Some(1_800_000_000));
-        assert_eq!(merged.last_success_timestamp, Some(1_700_000_100));
-    }
 }
